Finalize test_core_gates_features
