+++ src/qibo/tests/conftest.py
@@ -17,7 +17,6 @@ INACTIVE_TESTS = {
     "qibo.tests.test_core_distcircuit_execution",
     "qibo.tests.test_core_distcircuit",
     "qibo.tests.test_core_distutils",
-    "qibo.tests.test_core_gates_features",
     "qibo.tests.test_core_hamiltonians_from_symbols",
     "qibo.tests.test_core_hamiltonians_symbolic",
     "qibo.tests.test_core_hamiltonians_trotter",
