+++ src/qibo/tests/test_measurements_collapse.py
@@ -177,3 +177,40 @@ def test_measurement_result_parameters_multiple_qubits(backend):
     if int(m[2].outcome()): # pragma: no cover
         target_state = backend.apply_gate(gates.RX(3, theta=np.pi / 3), target_state, 4)
     backend.assert_allclose(final_state, target_state)
+
+
+@pytest.mark.parametrize("nqubits,targets", [(5, [2, 4]), (6, [3, 5])])
+def test_measurement_collapse_distributed(backend, accelerators, nqubits, targets):
+    initial_state = random_state(nqubits)
+    c = models.Circuit(nqubits, accelerators)
+    output = c.add(gates.M(*targets, collapse=True))
+    result = backend.execute_circuit(c, np.copy(initial_state), nshots=1)
+    slicer = nqubits * [slice(None)]
+    outcomes = [r.outcome() for r in output]
+    for t, r in zip(targets, outcomes):
+        slicer[t] = int(r)
+    slicer = tuple(slicer)
+    initial_state = initial_state.reshape(nqubits * (2,))
+    target_state = np.zeros_like(initial_state)
+    target_state[slicer] = initial_state[slicer]
+    norm = (np.abs(target_state) ** 2).sum()
+    target_state = target_state.ravel() / np.sqrt(norm)
+    backend.assert_allclose(result[0], target_state)
+
+
+def test_collapse_after_measurement(backend):
+    qubits = [0, 2, 3]
+    c = models.Circuit(5)
+    c.add((gates.H(i) for i in range(5)))
+    output = c.add(gates.M(*qubits, collapse=True))
+    c.add((gates.H(i) for i in range(5)))
+    final_state = backend.execute_circuit(c, nshots=1)[0]
+
+    ct = models.Circuit(5)
+    bitstring = [r.outcome() for r in output]
+    for i, r in zip(qubits, bitstring):
+        if r:
+            ct.add(gates.X(i))
+    ct.add((gates.H(i) for i in qubits))
+    target_state = backend.execute_circuit(ct)
+    backend.assert_allclose(final_state, target_state, atol=1e-15)
